# libs/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trader API",
    "summary": "Исследовательская платформа по рыночным данным",
    "description": "REST API платформы. Документация: `/docs` (Swagger UI), `/redoc`, схема — `/openapi.json`.\n\n**WebSocket** `/ws/jobs/{job_id}` (в OpenAPI не описывается): шлёт JSON в формате `JobOut` при каждом изменении задачи и закрывается по её завершении; код закрытия 4404 — задачи нет.",
    "version": "0.1.0"
  },
  "paths": {
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Поставить задачу в очередь",
        "description": "Задачу подхватит worker; ход выполнения — `GET /jobs/{id}` или WebSocket.",
        "operationId": "createJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Список задач",
        "description": "Новые задачи первыми; фильтры по статусу и типу.",
        "operationId": "listJobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "title": "Response Listjobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/engine-runs": {
      "get": {
        "tags": [
          "engines"
        ],
        "summary": "Прогоны движков",
        "description": "Новые сверху. Фильтры по движку, инструменту и таймфрейму.",
        "operationId": "listEngineRuns",
        "parameters": [
          {
            "name": "engine",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Имя движка",
              "title": "Engine"
            },
            "description": "Имя движка"
          },
          {
            "name": "instrument_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instrument Id"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timeframe"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngineRunOut"
                  },
                  "title": "Response Listengineruns"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/engine-runs/{run_id}": {
      "get": {
        "tags": [
          "engines"
        ],
        "summary": "Прогон движка",
        "operationId": "getEngineRun",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineRunOut"
                }
              }
            }
          },
          "404": {
            "description": "Прогон не найден"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/engine-runs/{run_id}/events": {
      "get": {
        "tags": [
          "engines"
        ],
        "summary": "События прогона (в том числе as-of)",
        "description": "`as_of` — срез знания: только события с `available_at <= as_of`. `view=history` — все события по порядку (включая пересмотры и отмены), `view=current` — актуальная картина на `as_of`: последняя версия каждой цепочки, отменённые исключены. `after_seq` — только события после этого номера (для догрузки).",
        "operationId": "listEngineEvents",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Типы событий",
              "title": "Kind"
            },
            "description": "Типы событий"
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "history",
                "current"
              ],
              "type": "string",
              "default": "history",
              "title": "View"
            }
          },
          {
            "name": "after_seq",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngineEventOut"
                  },
                  "title": "Response Listengineevents"
                }
              }
            }
          },
          "404": {
            "description": "Прогон не найден"
          },
          "422": {
            "description": "Слишком много событий: сузьте выборку"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EngineEventOut": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "detected",
              "confirmed",
              "revised",
              "invalidated"
            ],
            "title": "Status"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At",
            "description": "С какого момента событие известно"
          },
          "revises": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revises",
            "description": "seq пересматриваемого события"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "kind",
          "status",
          "payload",
          "detected_at",
          "confirmed_at",
          "available_at",
          "revises"
        ],
        "title": "EngineEventOut"
      },
...
      "EngineRunOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "algorithm_version": {
            "type": "integer",
            "title": "Algorithm Version"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "params_hash": {
            "type": "string",
            "title": "Params Hash"
          },
          "instrument_id": {
            "type": "integer",
            "title": "Instrument Id"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe"
          },
          "bars_processed": {
            "type": "integer",
            "title": "Bars Processed"
          },
          "last_close_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Close Time"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "engine",
          "algorithm_version",
          "params",
          "params_hash",
          "instrument_id",
          "timeframe",
          "bars_processed",
          "last_close_time",
          "event_count",
          "created_at",
          "updated_at"
        ],
        "title": "EngineRunOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreate": {
        "properties": {
          "type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9_.-]+$",
            "title": "Type",
            "description": "Тип задачи: `import.file`, `iss.sync_root`, `import.iss`, `aggregate.contract`, `demo.sleep`. Неизвестный тип worker завершит ошибкой."
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "description": "Параметры, зависят от типа задачи."
          },
          "max_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Attempts",
            "description": "Сколько раз повторять при сбое worker'а.",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "JobCreate",
        "description": "Новая фоновая задача.",
        "examples": [
          {
            "params": {
              "contract_id": 12
            },
            "type": "aggregate.contract"
          },
          {
            "params": {
              "contract_id": 12,
              "from": "2020-11-01",
              "till": "2020-11-30"
            },
            "type": "import.iss"
          }
        ]
      },
      "JobOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "progress_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Message"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "params",
          "status",
          "progress",
          "progress_message",
          "result",
          "error",
          "attempts",
          "max_attempts",
          "cancel_requested",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobOut",
        "description": "Состояние задачи. Статусы: queued, running, succeeded, failed, cancelled."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}